* fixed SKIN::text_y_offset not being scaled * fixed text_y_offset for modern skin

+++ trunk/src/gfx/gui/skin.h
@@ -72,7 +72,7 @@ namespace TA3D
         String prefix;
 
         //! offset to display text at the right place
-        int text_y_offset;
+        float text_y_offset;
 
     }; // class SKIN
